+++ app/src/main/java/org/breezyweather/sources/accu/json/AccuClimoSummaryResult.kt
@@ -15,22 +15,14 @@
  * along with Breezy Weather. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package org.breezyweather.db.entities
+package org.breezyweather.sources.accu.json
 
-import io.objectbox.annotation.Entity
-import io.objectbox.annotation.Id
-import java.util.Date
+import kotlinx.serialization.Serializable
 
 /**
- * History entity.
- *
- * [History].
+ * Accu climo summary
  */
-@Entity
-data class HistoryEntity(
-    @field:Id var id: Long = 0,
-    var formattedId: String,
-    var date: Date,
-    var daytimeTemperature: Float? = null,
-    var nighttimeTemperature: Float? = null
+@Serializable
+data class AccuClimoSummaryResult(
+    val Normals: AccuClimoNormals? = null
 )