+++ app/src/main/java/org/breezyweather/sources/openmeteo/OpenMeteoService.kt
@@ -169,6 +169,7 @@ class OpenMeteoService @Inject constructor(
     override val allergenAttribution = airQualityAttribution
     override val minutelyAttribution = null
     override val alertAttribution = null
+    override val normalsAttribution = null
 
     override fun requestSecondaryWeather(
         context: Context, location: Location,