+++ app/src/main/java/org/breezyweather/sources/openweather/OpenWeatherService.kt
@@ -123,6 +123,7 @@ class OpenWeatherService @Inject constructor(
     override val allergenAttribution = null
     override val minutelyAttribution = weatherAttribution
     override val alertAttribution = weatherAttribution
+    override val normalsAttribution = null
 
     override fun requestSecondaryWeather(
         context: Context, location: Location,